Move facsimile upload to tools/facsimiles.py
